+++ src/MicroElements.Swashbuckle.NodaTime/SwaggerGenOptionsExtensions.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Linq;
+using System.Text.Json;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.OpenApi.Models;
 using Newtonsoft.Json;
 using NodaTime;
 using NodaTime.Serialization.JsonNet;
@@ -15,38 +17,65 @@ namespace MicroElements.Swashbuckle.NodaTime
     {
         /// <summary>
         /// Configures swagger to use NodaTime types.
+        /// Uses NewtonsoftJson for serialization aspects.
         /// </summary>
-        /// <param name="config">Options to configure swagger.</param>
-        public static void ConfigureForNodaTime(this SwaggerGenOptions config)
+        /// <param name="config">SwaggerGenOptions.</param>
+        /// <param name="serializerSettings">Optional serializer settings.</param>
+        /// <param name="configureSerializerSettings">Optional action to configure serializerSettings.</param>
+        /// <param name="dateTimeZoneProvider">Optional DateTimeZoneProviders.</param>
+        public static void ConfigureForNodaTime(
+            this SwaggerGenOptions config,
+            JsonSerializerSettings serializerSettings = null,
+            Action<JsonSerializerSettings> configureSerializerSettings = null,
+            IDateTimeZoneProvider dateTimeZoneProvider = null)
         {
-            var serializerSettings = new JsonSerializerSettings();
-            config.ConfigureForNodaTime(serializerSettings);
-        }
+            serializerSettings = serializerSettings ?? new JsonSerializerSettings();
+            configureSerializerSettings?.Invoke(serializerSettings);
 
-        /// <summary>
-        /// Configures swagger to use NodaTime types.
-        /// </summary>
-        /// <param name="config">Options to configure swagger.</param>
-        /// <param name="initJsonSettings">Action to initialize jsonSettings.</param>
-        public static void ConfigureForNodaTime(this SwaggerGenOptions config, Action<JsonSerializerSettings> initJsonSettings)
-        {
-            var serializerSettings = new JsonSerializerSettings();
-            initJsonSettings(serializerSettings);
-            config.ConfigureForNodaTime(serializerSettings);
-        }
-
-        /// <summary>
-        /// Configures swagger to use NodaTime types.
-        /// </summary>
-        /// <param name="config">Options to configure swagger.</param>
-        /// <param name="serializerSettings">Settings to configure serialization.</param>
-        public static void ConfigureForNodaTime(this SwaggerGenOptions config, JsonSerializerSettings serializerSettings)
-        {
             bool isNodaConvertersRegistered = serializerSettings.Converters.Any(converter => converter is NodaConverterBase<Instant>);
             if (!isNodaConvertersRegistered)
-                serializerSettings.ConfigureForNodaTime(DateTimeZoneProviders.Tzdb);
+            {
+                serializerSettings.ConfigureForNodaTime(dateTimeZoneProvider ?? DateTimeZoneProviders.Tzdb);
+            }
 
-            Schemas schemas = new SchemasFactory(serializerSettings).CreateSchemas();
+            var nodaTimeSchemaSettings = serializerSettings.CreateNodaTimeSchemaSettingsForNewtonsoftJson();
+            config.ConfigureForNodaTime(nodaTimeSchemaSettings);
+        }
+
+        /// <summary>
+        /// Configures swagger to use NodaTime types.
+        /// Uses System.Text.Json for serialization aspects.
+        /// </summary>
+        /// <param name="config">SwaggerGenOptions.</param>
+        /// <param name="jsonSerializerOptions">Optional serializer options.</param>
+        /// <param name="configureSerializerOptions">Optional action to configure jsonSerializerOptions.</param>
+        /// <param name="dateTimeZoneProvider">Optional DateTimeZoneProviders.</param>
+        public static void ConfigureForNodaTimeWithSystemTextJson(
+            this SwaggerGenOptions config,
+            JsonSerializerOptions jsonSerializerOptions = null,
+            Action<JsonSerializerOptions> configureSerializerOptions = null,
+            IDateTimeZoneProvider dateTimeZoneProvider = null)
+        {
+            jsonSerializerOptions = jsonSerializerOptions ?? new JsonSerializerOptions();
+            configureSerializerOptions?.Invoke(jsonSerializerOptions);
+
+            global::NodaTime.Serialization.SystemTextJson.Extensions.ConfigureForNodaTime(jsonSerializerOptions,
+                dateTimeZoneProvider ?? DateTimeZoneProviders.Tzdb);
+
+            var nodaTimeSchemaSettings = jsonSerializerOptions.CreateNodaTimeSchemaSettingsForSystemTextJson();
+            config.ConfigureForNodaTime(nodaTimeSchemaSettings);
+        }
+
+        /// <summary>
+        /// Configures swagger to use NodaTime types.
+        /// </summary>
+        /// <param name="config">Options to configure swagger.</param>
+        /// <param name="nodaTimeSchemaSettings">Settings to configure serialization.</param>
+        public static void ConfigureForNodaTime(this SwaggerGenOptions config, NodaTimeSchemaSettings nodaTimeSchemaSettings)
+        {
+            config.ParameterFilter<NamingPolicyParameterFilter>(nodaTimeSchemaSettings);
+
+            Schemas schemas = new SchemasFactory(nodaTimeSchemaSettings).CreateSchemas();
             config.MapType<Instant>        (schemas.Instant);
             config.MapType<LocalDate>      (schemas.LocalDate);
             config.MapType<LocalTime>      (schemas.LocalTime);
@@ -69,6 +98,25 @@ namespace MicroElements.Swashbuckle.NodaTime
             config.MapType<Interval?>      (schemas.Interval);
             config.MapType<Offset?>        (schemas.Offset);
             config.MapType<Duration?>      (schemas.Duration);
+        }
+    }
+
+    /// <summary>
+    /// Resolves property name by <see cref="NodaTimeSchemaSettings"/>.
+    /// </summary>
+    internal class NamingPolicyParameterFilter : IParameterFilter
+    {
+        private readonly NodaTimeSchemaSettings _nodaTimeSchemaSettings;
+
+        public NamingPolicyParameterFilter(NodaTimeSchemaSettings nodaTimeSchemaSettings)
+        {
+            _nodaTimeSchemaSettings = nodaTimeSchemaSettings;
+        }
+
+        /// <inheritdoc />
+        public void Apply(OpenApiParameter parameter, ParameterFilterContext context)
+        {
+            parameter.Name = _nodaTimeSchemaSettings.ResolvePropertyName(parameter.Name);
         }
     }
 }